refactor: 將影片上傳元件中的工具函數移至 utils.ts，並刪除不再使用的 utils.ts 檔案，提升代碼結構的清晰度與可維護性。
